Updated ordering test so that the ordering test randomly selected how many operations to run instead of always running the same, fixed number of operations.
Note that this test relies on a bug in the Waldo compiler.  In its
call to

    logical_clear_switch_do_not_flush_clear_to_hardware

it calls clear on the flow table associated with a particular switch.
However, because the removes from calling .clear method do not
actually create set of updates to send to switch (see issue #6 in
github), it only clears the abstract representation of the flow table
(as we want in this, and only this, case).